Fixes for sanitizer.
